+++ plugit_proxy/views.py
@@ -431,6 +431,9 @@ def build_extra_headers(request, proxyMode, orgaMode, currentOrga):
         if 'REMOTE_ADDR' in request.META:
             things_to_add['remote-addr'] = request.META['REMOTE_ADDR']
 
+        if 'HTTP_X_FORWARDED_FOR' in request.META and getattr(settings, 'HONOR_X_FORWARDED_FOR'):
+            things_to_add['remote-addr'] = request.META['HTTP_X_FORWARDED_FOR']
+
         for meta_header, dest_header in [('HTTP_IF_NONE_MATCH', 'If-None-Match'), ('HTTP_ORIGIN', 'Origin'), ('HTTP_ACCESS_CONtROL_REQUEST_METHOD', 'Access-Control-Request-Method'), ('HTTP_ACCESS_CONTROL_REQUEST_HEADERS', 'Access-Control-Request-Headers')]:
             if meta_header in request.META:
                 things_to_add[dest_header] = request.META[meta_header]